AI_Core/requirements.txt
annotated-doc==0.0.4
tavily-python>=0.5.0
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-core==1.2.30
langchain-ollama==1.1.0
langchain-google-genai>=2.0.0
langgraph==1.1.6
langgraph-checkpoint==4.0.2
langgraph-checkpoint-mongodb==0.3.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.32
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
playwright==1.58.0
pydantic==2.13.1
pydantic_core==2.46.1
pymongo>=4.12,<4.16
pyee==13.0.1
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
starlette==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
xxhash==3.6.0
zstandard==0.25.0
